+++ src/Private/Assert-SNOWAuth.ps1
@@ -15,6 +15,7 @@ function Assert-SNOWAuth() {
         $CurrentTime = Get-Date
         $ExpiryTime = $script:SNOWAuth.Expires.AddSeconds(-$OauthExpiryBuffer)
         if($ExpiryTime -le $CurrentTime){
+            $ProxyAuth = $script:SNOWAuth.ProxyAuth
             #? Get a new token
             $Body = @{
                 grant_type="refresh_token"
@@ -22,7 +23,7 @@ function Assert-SNOWAuth() {
                 client_secret = [System.Net.NetworkCredential]::new('dummy', $script:SNOWAuth.ClientSecret).Password
                 refresh_token = $script:SNOWAuth.token.refresh_token
             }
-            $Token = Invoke-RestMethod -Method POST -uri "https://$($Script:SNOWAuth.Instance).service-now.com/oauth_token.do" -Body $Body -Verbose:$false
+            $Token = Invoke-RestMethod -Method POST -uri "https://$($Script:SNOWAuth.Instance).service-now.com/oauth_token.do" -Body $Body -Verbose:$false @ProxyAuth
 
             $script:SNOWAuth.token = $token
             $script:SNOWAuth.Expires = (get-date).AddSeconds($Token.expires_in)