+++ src/Public/sn_sc/Get-SNOWSCCart.ps1
@@ -15,11 +15,8 @@ function Get-SNOWSCCart {
     [CmdletBinding()]
     param()
     
-    Assert-SNOWAuth
     $URI = "https://$($Script:SNOWAuth.Instance).service-now.com/api/sn_sc/servicecatalog/cart"
-    $Headers = Get-AuthHeader
-
-    $Response = Invoke-RestMethod -Method "GET" -Uri $URI -Headers $Headers
+    $Response = Invoke-SNOWWebRequest -UseRestMethod -Uri $URI -Method GET
     if($Response){
         return $Response.Result
     }